+++ src/use-force-update.ts
@@ -1,8 +1,6 @@
-import { useState } from 'react';
+import * as React from 'react';
 
 export function useForceUpdate() {
-  const [value, set] = useState(true);
+  const [value, set] = React.useState(true);
   return () => set(!value);
 }
-
-export default useForceUpdate;