+++ src/use-media.ts
@@ -1,45 +1,19 @@
-/* eslint-disable react-hooks/rules-of-hooks */
-import { useEffect, useLayoutEffect, useState } from 'react';
-import json2mq from 'json2mq';
+import {
+  useMatchMedia,
+  useMatchMediaLayoutEffect,
+  QueryObject,
+} from './use-match-media';
 
-function createUseMedia(
-  effect: (
-    effect: React.EffectCallback,
-    deps?: React.DependencyList | undefined
-  ) => void
-) {
-  return function (
-    rawQuery: QueryObject | QueryObject[],
-    defaultState = false
-  ) {
-    let [state, setState] = useState(defaultState);
-    let query = json2mq(rawQuery);
+/**
+ * @alias useMatchMedia
+ * @deprecated
+ */
+export const useMedia = useMatchMedia;
 
-    effect(() => {
-      let mounted = true;
-      let mql = window.matchMedia(query);
-      mql.addListener(handleChange);
-      setState(mql.matches);
+/**
+ * @alias useMatchMediaLayoutEffect
+ * @deprecated
+ */
+export const useLayoutMedia = useMatchMediaLayoutEffect;
 
-      function handleChange() {
-        if (!mounted) return;
-        setState(mql.matches);
-      }
-
-      return () => {
-        mounted = false;
-        mql.removeListener(handleChange);
-      };
-    }, [query]);
-
-    return state;
-  };
-}
-
-export const useMedia = createUseMedia(useEffect);
-export const useLayoutMedia = createUseMedia(useLayoutEffect);
-export default useMedia;
-
-export type QueryObject = {
-  [property: string]: string | number | boolean;
-};
+export type { QueryObject };