+++ src/use-previous.ts
@@ -1,8 +1,8 @@
-import { useEffect, useRef } from 'react';
+import * as React from 'react';
 
 export function usePrevious<ValueType = any>(value: ValueType) {
-  const ref = useRef<ValueType | null>(null);
-  useEffect(() => {
+  const ref = React.useRef<ValueType | null>(null);
+  React.useEffect(() => {
     ref.current = value;
   }, [value]);
   return ref.current;