+++ src/use-update-effect.ts
@@ -1,18 +1,23 @@
-import { useRef, useEffect } from 'react';
+import * as React from 'react';
 
-export function useUpdateEffect(
-  effect: React.EffectCallback,
-  deps?: React.DependencyList
-) {
-  const mounted = useRef(false);
-  useEffect(() => {
-    if (mounted.current) {
-      effect();
-    } else {
-      mounted.current = true;
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, deps);
+export function createUseUpdateEffect(effect: typeof React.useEffect) {
+  return function useUpdateEffectHook(
+    effectCallback: React.EffectCallback,
+    deps?: React.DependencyList
+  ) {
+    const mounted = React.useRef(false);
+    effect(() => {
+      if (mounted.current) {
+        effectCallback();
+      } else {
+        mounted.current = true;
+      }
+      // eslint-disable-next-line react-hooks/exhaustive-deps
+    }, deps);
+  };
 }
 
-export default useUpdateEffect;
+export const useUpdateEffect = createUseUpdateEffect(React.useEffect);
+export const useUpdateLayoutEffect = createUseUpdateEffect(
+  React.useLayoutEffect
+);